refactor: Rename inspector/ to right-panel/, simplify migration
