+++ packages/client/src/right-panel/FilesTab.tsx
