+++ packages/client/src/right-panel/MetadataInspector.tsx
