+++ packages/client/src/right-panel/RightPanel.tsx
