+++ packages/client/src/right-panel/Section.tsx
