+++ packages/client/src/right-panel/agentDisplay.ts
